requirements.txt
flask
flask-cors
anthropic
voyageai
chromadb==0.5.3
beautifulsoup4
requests
python-dotenv
gunicorn
